requirements.txt
colorama==0.4.4
numpy==1.22.2
Pillow==9.0.1
